Add version requirement for hstlib to README.md
Thanks for what looks like a great tool (with a great name!). 

I noticed that the binaries require a more recent version of `hstlib` than are installed by default in some distros. This PR just makes
a note of the minimum version required and describes the error message that older versions will produce.